Suggested installation includes an explicit review step
While we did previously state "Read, then...", that's easier to ignore.
This brings the copy-pastable commands in line with the intent: to have
the user review the script before executing it. This also removes the
potentially confusing use of process substitution.

Note: The presence of `less` is assumed to avoid assumptions about
editors or the noise of handling that robustly (e.g. `${EDITOR:-vim}`)